Improve 'ellipsis uninstall' safety
Check for uncommitted files and changes before removing.

+++ src/ellipsis.bash
@@ -128,11 +128,22 @@ ellipsis.uninstall() {
         exit 1
     fi
 
-    pkg.env_up "$1"
-    pkg.run_hook "unlink"
-    pkg.run_hook "uninstall"
-    rm -rf "$PKG_PATH"
-    pkg.env_down
+    for package in "$@"; do
+        pkg.env_up "$package"
+
+        if pkg.run git.has_untracked || pkg.run git.has_changes; then
+            if ! utils.prompt "Uncommitted files or changes present, continue? [y/n]" "y"; then
+                pkg.env_down
+                continue
+            fi
+        fi
+
+        pkg.run_hook "unlink"
+        pkg.run_hook "uninstall"
+        rm -rf "$PKG_PATH"
+
+        pkg.env_down
+    done
 }
 
 # Re-link unlinked packages.
